+++ Magic/CHANGELOG.md
@@ -47,6 +47,7 @@
  - Add "equation" parameter to Command action, works just like Message action
  - Add "enable_vanilla_sounds" configuration option for those of you who *really* hate the new sound configs
  - Add `critical_probability` and `crticail_damage_multiplier` parameters to Damage action
+ - Added BossBar action
 
 # 8.2.3